Moves all compilation exceptions to the CompilerException class.
